python/requirements_dev.txt
# Copyright (c) 2025, NVIDIA CORPORATION.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

-r requirements.txt
black>=23.1.0
build>=0.10.0
isort>=5.12.0
mypy>=1.0.0
numpydoc
pydata-sphinx-theme
pylint
pytest
pytest-xdist
sphinx<6.0
twine>=4.0.0
